OCSADV-200: delete old vcs actions, port remaining actions to vcs2
