+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/viewer/SPViewerActions.java
@@ -17,9 +17,6 @@ public final class SPViewerActions {
     private final SPViewer viewer;
 
     // Individual actions
-    public final AbstractViewerAction vcsUpdateAction;
-    public final AbstractViewerAction vcsCommitAction;
-    public final AbstractViewerAction vcsSyncAction;
     public final AbstractViewerAction vcs2SyncAction;
     public final AbstractViewerAction syncAllAction;
     public final AbstractViewerAction conflictPrevAction;
@@ -132,9 +129,6 @@ public final class SPViewerActions {
         resolveConflictsAction = new ResolveConflictsAction(viewer);
 
         // VCS Actions
-        vcsUpdateAction = new VcsUpdateAction(viewer);
-        vcsCommitAction = new VcsCommitAction(viewer);
-        vcsSyncAction   = new VcsSyncAction(viewer);
         vcs2SyncAction  = new jsky.app.ot.vcs2.VcsSyncAction(viewer);
         syncAllAction   = new SyncAllAction(viewer);
