+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/viewer/SPViewerToolBar.java
@@ -64,7 +64,6 @@ public final class SPViewerToolBar extends JToolBar {
         add(new JPanel());
 
         add(viewer._actions.conflictNextAction);
-        add(viewer._actions.vcsSyncAction);
         add(viewer._actions.vcs2SyncAction);
         super.addSeparator();
     }
